refactor(restaurantUtils): move utility functions to shared module
Extract normalize, priceBucket and extractBestDiscountPercent functions from page component to restaurantUtils for better code organization and reusability
